Create extrac_suffix.py
Taking as input a file containing a lemma with its different grammatical forms, we extract the suffix from every grammatical form. Making a small change it could also extract the root and the prefix.
This Python script is based on the awk script I program for this purpose, it is available in the repository with the name "extrac_suffix.awk".